requirements.txt
Flask==1.0.2
gunicorn==19.5.0
python-socketio==2.0.0
eventlet==0.24.1